Refactor download
